requirements.txt
pynput
psutil
requests
Pillow
python-dotenv
mss
